% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF_OLS.R
\name{CF_OLS}
\alias{CF_OLS}
\title{CF_OLS.R}
\usage{
CF_OLS(
  Y,
  M,
  Covar,
  X,
  d,
  n,
  iter.max = 3,
  nsis = NULL,
  first.half = TRUE,
  seed = 2022,
  FDR = FALSE,
  FDRCutoff = 0.2,
  method = "iSIS"
)
}
\arguments{
\item{leftDmat}{n*(p+nknots+2) design matrix for left end of}
}
\value{
A list with the elements:
}
\description{
Calculate the test statistic and p-value for interval-censored competing risks SKAT.
}
\examples{
p0 <- 20
}
