% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Simulated data using Gompertz distribution}
\format{
A list of 4 studies with
\describe{
\item{X}{The time-to-event observed time.}
\item{Y}{The type of causes to the outcome: primary outcome (1) or competing event (2).}
\item{Y}{The type of causes to the outcome: primary outcome (1) or competing event (2).}
}
}
\usage{
exampleData
}
\description{
simulated time-to-event outcome and event type
}
\references{
\url{https://www.biorxiv.org/content/10.1101/2023.02.06.527391v1.abstract}
}
\keyword{datasets}
