% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_Meta.R
\name{R2_Meta}
\alias{R2_Meta}
\title{R2_Meta.R}
\usage{
R2_Meta(
  Effects,
  Study = NULL,
  SE,
  Method = c("Fixed", "REML", "PM", "DL", "ML")
)
}
\arguments{
\item{Effects}{Estimated R2M for each study.}

\item{Study}{An optional vector with study labels (Default = NULL).}

\item{SE}{Standard error of estimate for each study.}

\item{Method}{A character string indicating which method is used to estimate the between-study variance \eqn{\tau^2} and its square root \eqn{\tau}.}
}
\value{
A named vector with elements:
\describe{
\item{meta_effect}{Estimated R2-based total mediation effect using meta-analysis.}
\item{meta_CI_lower}{Lower bound of the 95\% confidence interval.}
\item{meta_CI_upper}{Upper bound of the 95\% confidence interval.}
\item{Q}{Statistic used to test for heterogeneity among the studies included in the analysis.}
\item{Q_pval}{The p-value associated with the statistic.}
\item{Method}{Method is used to estimate the between-study variance in random effects model or fixed effects model.}
}
}
\description{
A novel meta-analysis framework of the R-squared (R2)-based mediation effect estimation for high-dimensional omics mediators.
}
\examples{
Study <- c(1,2,3,4,5)
Effects <- c(0.1,0.2,0.3,0.4,1.5)
SE <- c(0.01, 0.02, 0.01, 0.08, 0.1)
Method <- "Fixed"
MetaR2M::R2_Meta(Effects=Effects, Study=Study, SE=SE, Method=Method)
Method <- "DL"
MetaR2M::R2_Meta(Effects=Effects, Study=Study, SE=SE, Method=Method)
}
