% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test.R}
\usage{
test(seed = 1)
}
\arguments{
\item{leftDmat}{n*(p+nknots+2) design matrix for left end of}
}
\value{
A list with the elements:
}
\description{
Calculate the test statistic and p-value for interval-censored competing risks SKAT.
}
\examples{
p0 <- 20
}
