% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF_OLS.R
\name{CF_OLS}
\alias{CF_OLS}
\title{CF_OLS.R}
\usage{
CF_OLS(
  Y,
  M,
  Covar,
  X,
  iter.max = 3,
  nsis = NULL,
  seed = 2024,
  FDR = FALSE,
  FDRCutoff = 0.2,
  method = "iSIS"
)
}
\arguments{
\item{Y}{Outcome input.}

\item{M}{Mediator(s) input.}

\item{Covar}{Covariates input.}

\item{X}{Exposure input.}

\item{iter.max}{Maximum number of iterations for (i)SIS and its variants (Default = 3).}

\item{nsis}{Number of pedictors recuited by (I)SIS (Default = NULL).}

\item{seed}{Random seed used for sample splitting, and cross-fitting of two subsamples (If the seed = NULL, the data will be evenly split into two halves).}

\item{FDR}{Indicator to perform FDR control.}

\item{FDRCutoff}{Cut-off point for FDR control.}

\item{method}{Methods for the mediators selection: iSIS or HDMT (Default = iSIS).}
}
\value{
A list with the elements:
\item{R2M}{Estimated R2-based total mediation effect.}
\item{SE}{Asymptotic standard errors for R2M.}
\item{CI_width}{Half width of the 95\% confidence interval.}
\item{CI_lower}{Lower bound of the 95\% confidence interval.}
\item{CI_upper}{Upper bound of the 95\% confidence interval.}
\item{p1}{Number of selected mediators in subsample 1.}
\item{p2}{Number of selected mediators in subsample 2.}
\item{R2M}{Estimated R2-based total mediation effect.}
\item{R2M}{Estimated R2-based total mediation effect.}
}
\description{
A more efficient two-stage cross-fitted estimation procedure for the R-squared measure in the presence of high-dimensional omics mediators.
}
\examples{
p0 <- 20
}
