% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Simulated data for tutorial}
\format{
A list of 4 studies with different sample sizes and each of them contains:
\describe{
\item{X}{The exposure mimics age}
\item{Y}{The outcome mimics systolic blood pressure}
\item{M}{The mediators}
}
}
\usage{
exampleData
}
\description{
simulated 4 independent studies for meta-analysis
}
\references{
\url{https://www.biorxiv.org/content/10.1101/2023.02.06.527391v1.abstract}
}
\keyword{datasets}
