% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_Meta.R
\name{R2_Meta}
\alias{R2_Meta}
\title{R2_Meta.R}
\usage{
R2_Meta(Effects, Study = NULL, SE, Method = "fixed")
}
\arguments{
\item{Effects}{Estimated R2M for each study.}

\item{Study}{Study name (Default = NULL).}

\item{SE}{Asymptotic standard errors for each study.}

\item{Method}{Methods for the meta-analysis.}
}
\value{
A list with the three elements:
\itemize{
\item A named vector with elements:
\describe{
\item{R2M}{Estimated R2-based total mediation effect.}
\item{SE}{Asymptotic standard errors for R2M.}
\item{CI_width}{Half width of the 95\% confidence interval.}
\item{CI_lower}{Lower bound of the 95\% confidence interval.}
\item{CI_upper}{Upper bound of the 95\% confidence interval.}
\item{p1}{Number of selected mediators in subsample 1.}
\item{p2}{Number of selected mediators in subsample 2.}
\item{R_YX}{Estimated R2 between X and Y.}
\item{R_YM}{Estimated R2 between M and Y.}
\item{R_YXM}{Estimated R2 between XM and Y.}
\item{SOS}{Shared Over Simple measure, which is R2M/R_YX.}
\item{SOS_CI_lower}{Lower bound of SOS.}
\item{SOS_CI_upper}{Upper bound of SOS.}
\item{Time}{Computational time (in minutes).}
\item{SampleSize}{Sample size used in the analysis}
\item{NumMeds}{Number of total mediators for the input.}
}
\item \code{m1}{Selected mediators in subsample 1.}
\item \code{m2}{Selected mediators in subsample 2.}
}
}
\description{
A novel meta-analysis framework of the R-squared (R2)-based mediation effect estimation for high-dimensional omics mediators.
}
\examples{
Study <- c(1,2,3,4,5)
Effects <- c(0.1,0.2,0.3,0.4,1.5)
SE <- c(0.01, 0.02, 0.01, 0.08, 0.1)
Method <- "fixed"
MetaR2M::R2_Meta(Effects=Effects, Study=Study, SE=SE, Method=Method)
}
